working on screen 2

+++ BankManagement_WPF/Model/UserDetail.cs
@@ -17,8 +17,8 @@ namespace BankManagement_WPF.Model
         public string State { get; set; }
         public string Country { get; set; }
         public string Email { get; set; }
-        public int PAN { get; set; }
-        public int Contact { get; set; }
+        public long PAN { get; set; }
+        public long Contact { get; set; }
         public DateTime DOB { get; set; }
         public string AccountType { get; set; }
