+++ LoginSecurity/Models/Domains/UserDetail.cs
@@ -18,8 +18,8 @@ namespace LoginSecurity.Models.Domains
         public string State { get; set; }
         public string Country { get; set; }
         public string Email { get; set; }
-        public int PAN { get; set; }
-        public int Contact { get; set; }
+        public long PAN { get; set; }
+        public long Contact { get; set; }
         public DateTime DOB { get; set; }
         public string AccountType { get; set; }
